# Fix SCCache closing races