--- a/src/hotspot/share/code/SCCache.cpp
+++ b/src/hotspot/share/code/SCCache.cpp
@@ -711,7 +711,7 @@ bool SCCHeader::verify_config(const char* cache_path, uint load_size) const {
   return true;
 }
 
-volatile int SCCache::_reading_nmethod = 0;
+volatile int SCCache::_nmethod_readers = 0;
 
 SCCache::~SCCache() {
   if (_closing) {
@@ -722,13 +722,7 @@ SCCache::~SCCache() {
   _closing = true;
   if (_for_read) {
     // Wait for all load_nmethod() finish.
-    // This is still racy: we cannot guarantee no one started reading
-    // nmethods after we closed the cache. We need to bail from readers
-    // when they detect the cache is closed. TODO: Fix this properly.
-    SpinYield spin;
-    while (Atomic::load(&_reading_nmethod) > 0) {
-      spin.wait();
-    }
+    wait_for_no_nmethod_readers();
   }
   // Prevent writing code into cache while we are closing it.
   // This lock held by ciEnv::register_method() which calls store_nmethod().
@@ -2885,6 +2879,11 @@ bool SCCache::load_nmethod(ciEnv* env, ciMethod* target, int entry_bci, Abstract
                            (entry->ignore_decompile() ? ", ignore_decomp" : ""));
   }
   ReadingMark rdmk;
+  if (rdmk.failed()) {
+    // Cache is closed, cannot touch anything.
+    return false;
+  }
+
   SCCReader reader(cache, entry, task);
   bool success = reader.compile(env, target, entry_bci, compiler);
   if (success) {
@@ -3378,6 +3377,10 @@ void SCCache::print_statistics_on(outputStream* st) {
   SCCache* cache = open_for_read();
   if (cache != nullptr) {
     ReadingMark rdmk;
+    if (rdmk.failed()) {
+      // Cache is closed, cannot touch anything.
+      return;
+    }
 
     uint count = cache->_load_header->entries_count();
     uint* search_entries = (uint*)cache->addr(cache->_load_header->entries_offset()); // [id, index]
@@ -3429,6 +3432,10 @@ void SCCache::print_on(outputStream* st) {
   SCCache* cache = open_for_read();
   if (cache != nullptr) {
     ReadingMark rdmk;
+    if (rdmk.failed()) {
+      // Cache is closed, cannot touch anything.
+      return;
+    }
 
     uint count = cache->_load_header->entries_count();
     uint* search_entries = (uint*)cache->addr(cache->_load_header->entries_offset()); // [id, index]
@@ -4182,3 +4189,57 @@ address AOTRuntimeConstants::_field_addresses_list[] = {
   card_shift_address(),
   nullptr
 };
+
+
+void SCCache::wait_for_no_nmethod_readers() {
+  while (true) {
+    int cur = Atomic::load(&_nmethod_readers);
+    int upd = -(cur + 1);
+    if (cur >= 0 && Atomic::cmpxchg(&_nmethod_readers, cur, upd) == cur) {
+      // Success, no new readers should appear.
+      break;
+    }
+  }
+
+  // Now wait for all readers to leave.
+  SpinYield w;
+  while (Atomic::load(&_nmethod_readers) != -1) {
+    w.wait();
+  }
+}
+
+SCCache::ReadingMark::ReadingMark() {
+  while (true) {
+    int cur = Atomic::load(&_nmethod_readers);
+    if (cur < 0) {
+      // Cache is already closed, cannot proceed.
+      _failed = true;
+      return;
+    }
+    if (Atomic::cmpxchg(&_nmethod_readers, cur, cur + 1) == cur) {
+      // Successfully recorded ourselves as entered.
+      _failed = false;
+      return;
+    }
+  }
+}
+
+SCCache::ReadingMark::~ReadingMark() {
+  if (_failed) {
+    return;
+  }
+  while (true) {
+    int cur = Atomic::load(&_nmethod_readers);
+    if (cur > 0) {
+      // Cache is open, we are counting down towards 0.
+      if (Atomic::cmpxchg(&_nmethod_readers, cur, cur - 1) == cur) {
+        return;
+      }
+    } else {
+      // Cache is closed, we are counting up towards -1.
+      if (Atomic::cmpxchg(&_nmethod_readers, cur, cur + 1) == cur) {
+        return;
+      }
+    }
+  }
+}