@shipilev 2024-09-04
@iklam, @veresov -- you might want to take a look. This blocks performance testing, because current `-XX:CacheDataStore` tests crash for me here a lot.