@shipilev 2024-09-05
> Please, use [premain] prefix in subjects which we use in JBS for our bugs. And Ioi filed SKARA request to automatically send our PR requests to Leyden-dev mailing list because I missed this PR.

Thanks! I see we do not prefix commit messages with `[premain]` when they do not have the associated JBS entries. So I'll skip this one, and see if we can integrated without JBS entry.

/integrate